Flyttet funksjon til separat fil
